requirements.txt
appdirs==1.4.3
Babel==2.4.0
cliff==2.6.0
cmd2==0.7.0
debtcollector==1.13.0
deprecation==1.0
funcsigs==1.0.2
functools32==3.2.3.post2
iso8601==0.1.11
jsonpatch==1.15
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.19.0
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.5
openstacksdk==0.9.15
os-client-config==1.26.0
osc-lib==1.5.0
oslo.config==4.0.0
oslo.i18n==3.15.0
oslo.serialization==2.18.0
oslo.utils==3.25.0
packaging==16.8
pbr==3.0.0
positional==1.1.1
prettytable==0.7.2
pyparsing==2.2.0
python-cinderclient==2.0.1
python-glanceclient==2.6.0
python-keystoneclient==3.10.0
python-novaclient==7.1
python-openstackclient==3.9.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
requestsexceptions==1.2.0
rfc3986==0.4.1
simplejson==3.10.0
six==1.10.0
stevedore==1.21.0
unicodecsv==0.14.1
warlock==1.2.0
wrapt==1.10.10
